Change the blah_blah_blah_md_header `index_length` member from `uint64_t` to `size_t` because it describes the size of an in-core structure as well as an on-disk one, and `size_t` is wide enough to store the size of any in-core structure, while `uint64_t` may be much too wide.  Check that `index_length` is no more than SIZE_MAX after we read it.

+++ src/H5FDprivate.h
@@ -200,7 +200,7 @@ typedef struct H5FD_vfd_swmr_md_header {
         uint32_t fs_page_size;
         uint64_t tick_num;
         uint64_t index_offset;
-        uint64_t index_length;
+        size_t index_length;
 } H5FD_vfd_swmr_md_header;
 
 static inline H5FD_vfd_swmr_idx_entry_t *
